**DO NOT SQUASH, MERGE THE COMMITS AS IS.**


| GroupId | ReleaseVersion | ReleaseSHA | ReleaseBuildId | ReleaseDate |
| --- | --- | --- | --- | --- |
| androidx.compose.animation | 1.11.0-alpha01 |
deb96499dfe95073f5c1215c1287787683cb1e92 | 14504571 | 12/3/2025 |
| androidx.compose.foundation | 1.11.0-alpha01 |
deb96499dfe95073f5c1215c1287787683cb1e92 | 14504571 | 12/3/2025 |
| androidx.compose.material | 1.11.0-alpha01 |
deb96499dfe95073f5c1215c1287787683cb1e92 | 14504571 | 12/3/2025 |
| androidx.compose.runtime | 1.11.0-alpha01 |
deb96499dfe95073f5c1215c1287787683cb1e92 | 14504571 | 12/3/2025 |
| androidx.compose.ui | 1.11.0-alpha01 |
deb96499dfe95073f5c1215c1287787683cb1e92 | 14504571 | 12/3/2025 |

Fixes https://youtrack.jetbrains.com/issue/CMP-9378

Follow up tasks:
- https://youtrack.jetbrains.com/issue/CMP-9385
- https://youtrack.jetbrains.com/issue/CMP-1284 (created much earlier)

## Testing
N/A

## Release Notes
N/A

+++ compose/ui/ui-inspection/src/androidTest/java/androidx/compose/ui/inspection/LazyColumnTest.kt
@@ -87,7 +87,7 @@ class LazyColumnTest {
 
     @Test
     fun rowsInOrder(): Unit = runBlocking {
-        rootId = getGlobalWindowViews().map { it.uniqueDrawingId }.single().toLong()
+        rootId = getGlobalWindowViews().map { it.uniqueDrawingId }.single()
         textComponentsInOrder(0)
         rule.onNode(hasScrollAction()).performScrollToIndex(30)
         rule.waitForIdle()
@@ -103,7 +103,7 @@ class LazyColumnTest {
     @Test
     fun testDrawModifierNodes() = runBlocking {
         generation++
-        rootId = getGlobalWindowViews().map { it.uniqueDrawingId }.single().toLong()
+        rootId = getGlobalWindowViews().map { it.uniqueDrawingId }.single()
         val checks = createAllParametersChecks(inspectorTester, rootId, generation)
         val withChildDrawModifiers =
             checks.composableResponse
